@@ -44,14 +44,14 @@
     "compression": "~1.7.4",
     "config": "~3.3.6",
     "date-fns": "~2.21.3",
-    "dotenv": "~9.0.2",
+    "dotenv": "~10.0.0",
     "express-rate-limit": "~5.2.6",
     "handlebars": "~4.7.7",
     "helmet": "~4.6.0",
     "morgan": "~1.10.0",
     "nestjs-typeorm-paginate": "~2.5.1",
-    "nodemailer": "~6.6.0",
-    "normalize-url": "~6.0.0",
+    "nodemailer": "~6.6.1",
+    "normalize-url": "~6.0.1",
     "openid": "~2.0.8",
     "passport": "~0.4.1",
     "passport-jwt": "~4.0.0",
@@ -59,7 +59,7 @@
     "pg-error-enum": "~0.5.0",
     "reflect-metadata": "~0.1.13",
     "rimraf": "~3.0.2",
-    "rxjs": "~6.6.7",
+    "rxjs": "~7.1.0",
     "st-utils": "~1.7.0",
     "swagger-ui-express": "~4.1.6",
     "typeorm": "~0.2.32",
@@ -75,25 +75,25 @@
     "@types/bcryptjs": "~2.4.2",
     "@types/compression": "~1.7.0",
     "@types/config": "~0.0.38",
-    "@types/express": "~4.17.11",
+    "@types/express": "~4.17.12",
     "@types/express-rate-limit": "~5.1.1",
     "@types/fs-extra": "~9.0.11",
     "@types/inline-css": "~3.0.0",
     "@types/jest": "~26.0.23",
     "@types/morgan": "~1.9.2",
     "@types/node": "~14.14.45",
-    "@types/nodemailer": "~6.4.1",
+    "@types/nodemailer": "~6.4.2",
     "@types/openid": "~2.0.1",
     "@types/passport-jwt": "~3.0.5",
     "@types/socket.io": "~2.1.13",
     "@types/supertest": "~2.0.11",
     "@types/uuid": "~8.3.0",
-    "@typescript-eslint/eslint-plugin": "~4.24.0",
-    "@typescript-eslint/parser": "~4.24.0",
+    "@typescript-eslint/eslint-plugin": "~4.25.0",
+    "@typescript-eslint/parser": "~4.25.0",
     "adm-zip": "~0.5.5",
-    "eslint": "~7.26.0",
+    "eslint": "~7.27.0",
     "eslint-config-prettier": "~8.3.0",
-    "eslint-plugin-import": "~2.23.2",
+    "eslint-plugin-import": "~2.23.3",
     "fs-extra": "~10.0.0",
     "jest": "~26.6.3",
     "ora": "~5.4.0",
@@ -103,7 +103,7 @@
     "ts-loader": "~8.1.0",
     "ts-node": "~9.1.1",
     "tsconfig-paths": "~3.9.0",
-    "typescript": "~4.2.4",
+    "typescript": "~4.3.2",
     "yargs": "~17.0.1"
   },
   "jest": {
